gym_backend/requirements.txt
# Core Django packages
Django==4.2.23
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
django-extensions==3.2.3  # Database shell plus and other utilities

# Caching and Performance
redis==5.0.1
django-redis==5.4.0
django-cache-machine==1.2.0

# Authentication and Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-axes==6.1.1  # Brute force protection

# Monitoring and Logging
django-extensions==3.2.3
sentry-sdk[django]==1.39.2
django-debug-toolbar==4.2.0

# Performance and Optimization
django-silk==5.0.4  # Profiling
gunicorn==21.2.0  # WSGI server
whitenoise==6.6.0  # Static file serving

# Database optimization
django-bulk-update==2.2.0
django-model-utils==4.3.1

# API Documentation
drf-spectacular==0.27.0

# Development
python-decouple==3.8  # Environment variables